Create a ResourceLoader for an EditorSession using its associated installation or instance
EditorApp creates the ResourceLoader for an EditorSession before creating it.
EditorApp finds the install or instance and calls getResourceLoader on it
MMCInstance now has getResourceLoader
MMCInstance's getResourceLoader returns a loader with all installed mods in its search path. Not sure whether they should go before or after the MC jar(s).  (todo: add installed mods for vanilla installs, for homecooked mod packs/installs?)
MultiMCInstall now has versions.
MCInstallGroup is now responsible for finding a v1.8 jar, and searches all installs and MMC installs for one.
ensureValidInstall checks for v1.8 and displays an appropriate dialog if it isn't found.

All of this will make it easier to hook in a "Minecraft Assets Pack" as a separate download, and makes it possible to load assets from any installed mods