Encoding pathnames with sys.getfilesystemencoding is always wrong.
On Windows this encodes with the current codepage, which often fails for characters not in the codepage.

On Windows, FS functions accept 'unicode' and encode as UTF-16 for WinAPI
On Mac, FS functions accept 'unicode' and encode as UTF-8 for FS API
On Linux, filenames are byte strings and encoding is only a convention. FS functions encode 'unicode' according to locale. On Linux, it is possible to have files whose names cannot be decoded in any character set!