Avoid creating world view when world list is offscreen.
Seems to fix the access violations, but that's not at all certain.
Hours of bisecting and single-stepping also blamed the simple import of
one of the UI files for the violations.

In any case, this avoids doubly creating the world view both when the
window is initialized and when it is shown on screen.